requirements.txt
pandas
flask
plot
plotly-express
psycopg2-binary